Revert "Move Collapsing Frame"
This reverts commit 8beb79cace94e5140fc417daaee5a0c4c3889440.
