+++ maskgen/collapsing_frame.py
